Restructure ml.models.
